Fix issues related to specifying path in `ALCF/*`
Previously, we had been using a (hacky) method for retrieving the absolute path
to `Megatron-DS-Benchmarking/ALCF/`, using bash to determine where we are,
and then manually appending the path to this location.

This is, to say the least, not _ideal_.

In this commit, we can explicitly (guarantee) the location by determining where
`megatron` is installed to, e.g.

`MEGATRON_DIR=$(dirname $(dirname $(python3 -c 'import megatron; print(megatron.__file__)')))`

and then `ALCF_DIR="${MEGATRON_DIR}/ALCF` is _guaranteed_ to point to the
correct location.